Выполнены простые шаги по заданию 5-го спринта
