+++ MovieQuiz/Models/QuizResultsViewModel.swift
@@ -1,0 +1,14 @@
+//
+//  QuizResultsViewModel.swift
+//  MovieQuiz
+//
+//  Created by Захар Панченко on 07.07.2024.
+//
+
+import Foundation
+
+struct QuizResultsViewModel {
+    let title: String
+    let text: String
+    let buttonText: String
+}
