+++ MovieQuiz/Models/QuizStepViewModel.swift
@@ -1,0 +1,14 @@
+//
+//  QuizStepViewModel.swift
+//  MovieQuiz
+//
+//  Created by Захар Панченко on 07.07.2024.
+//
+
+import UIKit
+
+struct QuizStepViewModel {
+    let image: UIImage
+    let question: String
+    let questionNumber: String
+}
